Refactor validations to account for uploaded images
Previously, new Movies were considered invalid if they did not contain a 'poster_image_url'. This commit adds a custom validation method 'has_image?' that checks for the presence of either a url or an uploaded file. If neither is found, an error is added to both the 'uploaded_image' field and the 'poster_image_url' field stating that at least one field must have an image.