Update views to display uploaded images
Adds an 'if' check to the image display section for movies 'index' and 'show' views. If a user uploaded image is present it's displayed (first choice), otherwise, if an image URL is provided then that image is displayed (second choice). If neither is present, nothing is displayed: given that there is a validation to make sure that each new/edited movie has an image, this should not happen, but this was added because some movies did not contain images during testing and that caused errors.