Move search functionality from Movie controller to Movie model as class method
Since the logic for the expanded search is more complex and can no longer be contained on one line, it makes more sense to turn it into a class method than to keep it in the controller.
The new Movie.search method acts as follows:
- Takes the search_params as an argument (defined in Movie controller)
- uses scopes to filter entries in the Movies table, such that the search query will build upon itself
to further narrow the results with each additional parameter

The search method will not be called if the search parameters are empty, which means the index page will continue to load normally (show all movies).