Abstract header to layout.html.erb
Remove the 'Rotten Mangoes' headline from each page and place it in the layout file. Also turn headline into a link back to the movies index view, so that the 'Back to all movies' links can be removed from the pages they were found on. This change was prompted by the search results page, which changes the movie list for the movies index view -- because the search results simply uses the index view with a filtered list of movies, there was no easy way to go 'back' to the full list.

Note that a couple pages had to have their headlines converted from <h1> to <h2> to make more sense with the new site-wide header.