Define method for sending goodbye email
Set standard settings for sending a goodbye email (confirmation of account deletion). The method takes a user as an argument. It assumes that the user has an 'email' attribute, and calls the mail() function to send a message to that email address with a default subject line - set here as 'Account deleted'.